Change dump_app_state signature to make it more flexible
dump_app_state now receives an arbitrary list of arguments that are then used
to update the current state of the application. The logic goes as:
- dump_app_state receives the state properties to update as **kwargs
- it reads the current state of the application from the on-disk JSON
- it updates the dictionnary from the **kwargs
- it stamps it with the current date
- and write the update state back to disk

This patch also name the set_pas_incr function properly. Previously this one was
named set_cycles as well, shadowing the proper set_cycles function